Add macro literal ident escape
PR #65 removed support for R7RS style literals in macro definitions.
These were a list of identifiers that were considered literals for the
whole macro body. This simplified our syntax and implementation and
allowed using keywords in their place.

However, there are a few operator/punctuation idents that we may want to
match on - e.g. `_`, `&` and `...`. These are technically idents so
right now matching is impossible as they'll be treated as variables.

However, R7RS already has a escape for `...` - `(... ...)`. This works
because it's illegal for the ellipsis to appear at the beginning of the
list. This extends that escape in two ways:

1. Allow arbitrary idents as the second member of the list, e.g.
   `(... &)`

2. Allow the escapes to appear in both the pattern and template and
   handle them in the matcher

The implementation is simple but the syntax is ugly. However, this is
intended to be used sparingly as an escape hatch and reuses existing
macro syntax.