Refactor spans and reporting
This was prompted by evaluating the `codespan` crate. This could
potentially replace `span.rs`, `loader.rs` and `reporting.rs` with
something much more mature and with possible Language Server Protocol
support in the future.

Jumping directly to the `codespan` crate isn't trivial. However, because
we've both been inspired by the Rust compiler a lot of the concepts and
names are similar. The interface of the `codespan` crate is generally
better thought out. This is an opportunity to steal some of those ideas
as standalone improvements while also making a potential switch to
`codespan` easier:

- Rename Span's `lo`/`hi` to `start`/`end`. There were already a few
  places calling it the latter.

- Make `Span` an opaque struct with a constructor and field accessors

- Rename `reporting::Level` to `reporting::Severity`

- More consistently use `u32` when dealing with byte indexes that are
  used to create or offset spans

- Leave a 1 byte padding between all source file spans. This helps
  prevent edge cases.