Replace SourceFileId with RcId
SourceFileId was basically tricking the borrow checker by stashing a
vector index and then pulling out of the vector when we needed the. We
internally ensured the invariant that we would never delete or replace a
vector item. However, we could still e.g. use `SourceFileId` on the
wrong `SourceLoader`.

This was inspired by `codespan` which uses `Arc` for its analogous
`FileMap`.