Always load RFI Arret types in to SourceLoader
We previously had a hack that would only load the Arret type as source
if there was an error. This was an optimisation to avoid bookkeeping for
every exported RFI function. However, we recently added `Span`s to our
polymorphic variables for better error reporting. This means that
certain errors can point back to RFI Arret types even if the RFI module
was loaded successfully.

We originally avoided this for performance reasons. However, we can
mitigate that with a few strategies:

- Make our `SourceLoader` use `Cow<'static>` so we don't need to copy
  the source

- Make the function name `&'static str` so we don't need to copy the
  function name

- Make the filename `Rc` so it can be shared between all funs in an RFI
  module