Cache parsed source in our loader
This allows us to cache the full parsed syntax AST instead of just the
source string. Our parsing is consistently near the top of our profiles
for our integration tests as every test program is reparsing the
standard library. This will only get worse as the standard library and
test suite grow.