Make RFI loader return Rc references
This adds an additional caching layer for loaded libraries in
`rfi::Loader` and then has it return `Rc`s to a library instead of its
exports. It's then up to `LoweringCtx` to track the libraries it got
back.

This gives no immediate benefit but it's a step in hoisting
`rfi::Loader` up to the same level as `SourceLoader` to cache RFI
libraries across programs.