Improve lock contention in SourceLoader
Right now we can only load a single source file at a time due to locking
our path cache. This includes the time it takes to actually read the
file from disk. In the case of our integration tests the first thing we
do is read one file per thread which will immediately block. This adds a
`load_path_uncached` which skips the path cache entirely which allows us
to read the file without taking a lock until we parse it.

This also uses a separate lock for the path cache and actual parsing
(which is needed to keep the Span offset consistent). This should mean
that loading the file from disk won't contend with e.g. `load_string`
for RFI functions.