Don't require original scope when expanding macros
Currently the meaning of "scope" in our lowering phase is a bit unusual.
All modules are actually in the same scope with different namespace IDs.
This is so when we expand macros we can look up the original binding
from the scope.

In addition to being strange this also makes it difficult to multithread
module lowering. This is because the scope needs to be "threaded"
through every module as its lowered.

Instead, change our macro linker to resolve template idents during our
link phase. This means we have everything we need during expansion
without needing to look anything up in the scope.

This makes it possible to build a scope for each module we lower and
completely discard it at the end. However, this is not implemented yet.