Basic DCE during type inference
Our dead code elimination predominantly happens during the MIR phase.
Because MIR uses partial evaluation and tracing of unused MIR ops it can
be much more precise and complete. However, MIR looks at the expression
tree when calculating a closure so it'd be useful to remove obviously
dead code earlier.

There are three places we can do dead code elimination:

1. When visiting non-terminal pure expressions in a `do`

2. When visiting all expressions in a `do` after a divergence

3. When using `keep_exprs_for_side_effects`. This is used when constant
   folding conds and type predicates.

In all cases we run type inference first to ensure the dead expressions
are well-typed.

This reduces the instruction count of our integration tests by 1.5%
presumably by avoiding to have to partially evaluate the dead code. This
wasn't the intention of early DCE but it's a nice bonus.