Slim down `value::ArretFun`
We had grown huge (176 bytes) which was bloating the size of
`value::Value`. Slim us down by moving everything but our closure in to
and `Rc`. The closure needs to be separate so it can be mutated during
garbage collection without breaking the whole `Rc`.

This brings us down to 40 bytes.