Get rid of TopPair, TopVector and TopList
`TopFoo` was similar to but a distinct type from `Foo<Any>`. This was
because our `Top` version participated in our system of macros that
handles our tagging while the `Foo<T: Boxed>` version was a generic type
that happened to have an identical data layout. There were methods
provided to switch between the two representations.

If that sounds complicated and confusing it's because it was. While it
would be manageable if this was isolated to the runtime it bled out in
to our public RFI interface. This means library authors would be exposed
to the multiple personalities of our generic types.

One of the reasons this wasn't done previously is that e.g. `Pair<T>`
needed to implement traits such as ABI encoding that we also wanted to
provide automatically for our macro defined tags. Trying to unify them
would cause overlapping implementations for e.g. `Pair<T>` and
`Pair<Any>`.

To deal with the trait problem we split `DirectTagged` in to
`ConstTagged`, `DistinctTagged` and `UniqueTagged`. See their
documentation for more details on their precise definition. We then give
`Vector` and `Pair` default type arguments to allow our tagging macros
to deal with them transparently.

This also merges more of the implementation between tagged unions and
`Any`. Tagged unions has a limited subset of `Any`s functionality. Now
they're at parity via a `define_supertype!` macro.