Document the `runtime::boxed` module
This is the majority of the public interface for RFI so it's useful to
document. This adds a `#[warn(missing_docs)]` to ensure we don't regress
on documentation. Other modules might grow the same warning as our
rustdocs are expanded.

There are a few semi-private things such as the garbage collector in
there as well but it was easier to document them than exclude them from
our doc warnings.