Cache synthetic funs during MIR
Whenever `=` or a type predicate is treated as a value we need to
generate a synthetic `ArretFun`. This is essentially a templated snippet
of HIR created on demand.

These weren't being cached which means every instance of a synthetic
function was being treated as unique and generating completely separate
thunks, native functions, etc. This also means that testing equality
between synthetic functions wasn't unreliable.

Fix by caching the `ArretFun`s and reusing instances in the same
`EvalHirCtx`