Treat all funs as inequal
With the different representations a function can take at runtime (boxed
thunk + an arbitrary number of callbacks) it's hard to reliably
determine if two functions are the same. This is complicated further by
the presence of closures. Previously we gave a best effort approach but
this wasn't consistent between compile time and runtime and could vary
with different optimisations, inlining, etc.

Punt this completely by treating it the same way as `NaN` - all function
values are inequal, even if it's literally the same var.