Fix occurrence typing of `(and)`
Occurrence typing of `(and)` appeared to work in the case of multiple
type conds being applied to the same var. For example:

```
(defn maybe-first ([lst (U (List & Any) false)]) -> (U Any false)
  (if (and (list? lst) (not (nil? lst)))
    (first lst)
    false))
```

This works because the type cond for `(nil?)` is generated in a context
where `(list?)` has already narrowed the type. This causes it to
generate a type cond containing the correct type.

This falls apart when `(and)` is referring to multiple variables:

```
(defn two-var-and-typing ([left Num] [right Num])
  (when (and (int? left) (float? right))
    (ann left Int)
    (ann right Float)))
```

In this case only the terminal `(float?)` has an effect and the earlier
type conds are lost. This is because except in one special case for
`(not)` the cond's from an `(if)`'s test are never propagated upwards.

The above example macro expands to this:

```
(if (if (int? left) (float? right) false)
  (do
    (ann left Int)
    (ann right Float))
  ())
```

Because the false branch of the terminal `(if)` is false if the `(if)`
evaluates to true the test must be true. This allows us to propagate the
type cond upwards in this case and its boolean mirror.

This was tricky with our existing type cond representation of having a
true/false type pair for each cond. In the above we can only propagate
the true half of the type cond; we can make no guarantees about what
happens if its value is false because that could've come from the true
branch.

This splits the type cond representation in to one entry each for the
true and false case. This allows us to propagate just the type conds
that we're allowed to and lets us avoid creating useless type cond
halves in some cases.

There is an additional level of trickiness in that we may now have
multiple type conds for the same var in some common circumstances. In
the very first example of a working `(and)` there would now be a less
specific type cond restricting it to a list and a second one to a
non-empty list. For now we just apply them in order as they should
become increasingly specific as we apply. However, we need to be careful
to restore the types in the reverse order to recover the original type.