Use unique names for equality regs
Right now one of the most common names for LLVM registers is
`int_equals`. This is because its used not only for integer equality but
also bools, chars, interned symbols and type tags. All of those ops
except for type tag equality already had a dedicated op. However, they
shared a common code path that used `int_equals`. This splits the code
paths using a helper function with the reg's name. This also adds a
`TypeTagEquals` op for consistency with the other typed equalities.