Treat pure as satisfying all purities
Much like impure is the top purity (analogous to `Any`) pure is the
bottom purity (analogous to `(U)`). We should consider pure as
satisfying any polymorphic purity.

In practice this comes up when using pure functions in a function with
polymorphic purity. Prior to this commit that was erroneously
disallowed. The added `eval-pass` test is testing this exact case.