Generalise (and) and (not) occurrence typing
The occurrence typing logic for `(and)` and `(not)` could be thought of
of special cases of using a cond branches' static types to make
assertions about the cond test's type. We can merge them to both
simplify the code and handle occurrence typing in more cases.

I can't think of a non-contrived example corresponding to the newly
covered cases. However, macros have a knack for producing convoluted
conds that confuse occurrence typing.