Add impure `(panic)` variant
`(panic)` is a pure function to allow usage for unexpected cases in pure
functions without forcing the whole function to impure. However, this
leads to a problem with our integration tests where we explicitly want
to evaluate the `(panic)` even if its result is technically unused. We
have a hack in `eval_hir` which tries to treat `(panic)` as semi-impure
to ensure it's kept for its side effects.

Unfortunately this only works due to our optimiser not being clever
enough. Many of the `(test-*)` helper integration test functions are
also being considered pure and the optimiser could easily remove their
application completely.

This approaches the problem from two angles:

1. It adds a `(panic!)` variant and updates `(assert-*)` to use the
   impure variant. The names of the assert macros are update to reflect
   their impurity.

2. It explicitly marks our test helper functions as impure

This also allows us to remove the `eval_hir` special case for `(panic)`.