Use `Void` ABI return type for Arret funs
Currently if an Arret fun returns the empty list we'll generate a
function returning a constant boxed pointer to `Nil`. This is correct
but we can use the `RetABIType::Void` to skip the return entirely. This
was already used by the RFI for Rust funs returning `Nil`.

It's likely that LLVM figures this out itself in most cases but this can
only make its job easier.