Fix for partial equality of identical regs
We have a fast path for equality when comparing the same register
against itself. This is intended for cases such as macro generated code
or partial evaluation have produced some version of `(= x x)`. This was
already handling float's partial equality (NaN != NaN) but it missed two
major cases:

1. Functions are very partially equal - they are never equal to any
   value, including themselves
2. Vectors and lists may contain partial equal values

This extends our identical reg test to cover the above. In addition we
teach `values_statically_equal` that a function reg is never equal to
anything else.