Add intrinsics for numeric comparisons
This generates native code for comparisons of 1 or more `Num` values
with known types. This misses a few cases where we can determine a
comparison is statically false (e.g. if a pair comparison is false or we
encounter a NaN). However, those should be fairly obscure cases that
LLVM will partially compensate for.