Fix applying FunThunk values at compile time
Commit 4c737ce started treating all `FunThunk` values as inequal. While
this works at the language level it broke MIR's mapping from `FunThunk`
to `Value` which prevents us from treating functions as first-class
values.

The easiest way to test this was to promote the application integration
test to `run-pass`. I want to merge these tests together anyway so this
is a step in the right direction.

This also changes an `unimplemented!` to a `panic!` if we can't find the
`FunThunk`. This should occur unless some invariant has been violated
such as RFI code producing its own `FunThunk`s.