Allow poly pure applications in impure functions
Right now we always add our outer function purity as evidence for the
application's purity. This is required for type safety in pure and
polymorphic purity outer functions to make sure the application isn't
more impure than the function allows. However, if the function is
explicitly impure this just forces all polymorphic purity applications
to be impure.

This is type safe but prevents our MIR from optimising the applications
(e.g. compile-time eval or elision).  This is unexpected and
inefficient. Fix by not adding fixed impurity as evidence during type
inference.

This also re-adds an optimisation for not building empty integration
tests. This fix allows the application integration test to be
completely optimised away.