Merge eval-pass and run-pass integration tests
The `eval-pass` tests were only supposed to evaluated at compile time
and not produce a binary. However, it's hard to test intrinsics with
eval because everything is const evaled by the standard library. This
caused us to have mirrored `eval-pass` and `run-pass` tests where the
`run-pass` was a cut down version using `black-box!`. This also caused
us to miss a severe compile-only bug (see e4abe56).

This merges the two tests together in to one type. The special
`optimise` tests are still implemented the same way as the former
`eval-pass` but there should be one less user visible test type to worry
about.