Refactor Expr spans
`hir::Expr` unconditionally contained a span. This led to the use of
duplicate or nonsense spans in some cases:

- `ExprKind::Lit` already had a span in its datum

- `ExprKind::RustFun` had a span on `rfi::Fun`

- `ExprKind::Do` was a pseudo-expr that only had a span in the case of a
  literal `(do)`

This removes the span from `Expr` and instead moves in on to the kinds
where it makes sense. This is part of the work to remove non-test uses
of `EMPTY_SPAN`.