More helpful message on first `cargo test`
Our compiler tests dynamically load the `stdlib` as that contains many
of the language's base functions needed to write integration tests.
Right now I can't find an easy way to express to Cargo that the `stdlib`
cdylib needs to be built before the compiler. This causes a `cargo test`
on a fresh checkout to fail. This is a bad look for new developers
playing with Arret.

Until I can fix the ordering properly add a message telling developers
what they need to do.