Switch to codespan for reporting
This replaces the `compiler::Error`/`compiler::Reportable` types and
reporting code with the equivalent implementation in `codespan`.

This is nearly a direct translation from our previous error
implementation. One thing we've lost is the carets under the source
snippet. This is because `codespan` (and Rust) typically have a separate
label message from the error message. In its absence the carets aren't
printed. We can gradually introduce proper label messages as separate
changes.

This also fixes our `error::Error` implementations to print a useful
message by reusing the diagnostic code.