Switch import filter names to use keywords
These don't function as normal identifiers - i.e. they're not bound in
the scope and are instead matched entirely on their name. This is also a
bit more consistent with what Clojure does visually although the
structure of our filters is still completely different.

This also renames `:except` to `:exclude` for consistency with Clojure
and to be clear there is no association with exceptions.