Add support for fixed record polymorphic vars
The biggest conceptual mismatch between our implementation of
polymorphic functions and polymorphic records is the concept of "fixed"
type variables.

Our type system is designed and tested around the assumption that
polymorphic variables must have more than one possible type. This
simplifies its logic and allows it to make stronger assertions about
those "fixed" polymorphic variables. As a result our `hir` detects these
cases and replaces the usage of the fixed variables with the known type.

Declaring polymorphic variables with one possible value might seem
useless. However, it's easy to see how this could happen with
macro-generated code that takes a type as input.

There is an additional complication around record types as the order of
the polymorphic parameters need to be preserved for the type constructor
we automatically build. Theoretically this should live entirely in the
`hir` but that would require `hir` to track additional metadata on top
of the record type and pipe that in to `str_for_ty_ref` etc.

This adds two more variants of `record::PolyParam` for fixed purity and
type variables. Currently these are ignored except inside
`str_for_ty_ref`.