Include usage in "expected symbol" error messages
There are many places in our syntax where we expect symbols. Simply
telling the developer that a symbol was expected may not be enough.
Instead, include a simple usage string that's used for the label in the
error message indicating what the symbol would be used for.

This also switches the order of the returned tuple from our helper to
`(span, ident)`. Everywhere else (function parameters, struct fields,
etc.) the span comes before the data it's spanning by convention. The
flipped order was jarring compared to the rest of the codebase.