Import ty::Ty as Ty
We already do this for `purity::Purity` which is the analogous enum for
purities. This is a bit confusing as we now have `ty::List` and
`Ty::List` but the latter only exists to be converted to the former.

The main motivation for this is to make programmatically specifying
types less verbose, especially in tests. The medium term goal is to
reduce or eliminate the usage of type strings and `poly_for_str` in our
tests. That would reduce the interdependency between our HIR and type
system.