Add a top record type
This makes the new record type hierarchy:

```
Any > Top Record > Record Class > Record Instance
```

This is mainly motivated by the need to be able to convert back and
forth between Arret types and ABI types. For example,
`ops_for_arret_fun` needs to know which Arret type corresponds to a
boxed record.

This could've been worked around with record-specific checks but it's
easier to just add top records to the type system.

This is also exposed to Arret with the `Record` type and the `record?`
type predicate. Unlike the record class type predicates this is fully
supported by MIR and codegen because it's a simple type tag check.