Support for record class type predicates
This is the first type predicate that can't be decided solely by their
type tags; the record class ID basically acts as a secondary type tag.
We need separate paths for the existing "type tag" predicates and the
new record class type predicates.

After making an effort to evaluate the type predicate statically the
logic is:

1. Determine if its possible for the value to *not* be a record. If so,
   check its type tag. If it's not a record then return false.

2. Cast the value to a boxed record.

3. Compare the value's record class ID to our test record class ID.