Add boxed::{PartialEqInHeap, HashInHeap}
Records require additional information about their field layout to
implement equality, hashing and garbage collection. For example, the
garbage collector needs to know the offset of every pointer to a boxed
value so it can trace and update them.

There are a few ways this could be done:

1. Inline the layout information inside each record instance. This can
   take an arbitrary amount of storage.

2. Include a pointer to the layout information. This needs to be 8 bytes
   on 64bit platforms.

3. Include an integer that can be used to lookup the layout information.
   This can easily be 4 bytes on all platforms.

Option 1 is clearly too much storage overhead. However, option 2 is much
more expensive than it seems due to our box layout. The numbers below
assume natural alignment and 64bit pointers which are typical for our
target platforms.

Right now, all boxes start with a 2 byte header. Including a 64bit
pointer would require 6 bytes of padding before the 8 byte pointer to
keep it aligned. This means we'll use all of a 16 byte box for metadata
requiring a 32 byte box or out-of-line data to actually store any
information.

However, a 32bit class ID will only require 2 bytes of padding. This
comes to a total of 8 bytes which leaves space for another 8 bytes of
data. This can store at least one field of any value in a 16 byte box.

The downside of this is we need access to our field layout lookup table.
This is similar to how we need the interned symbol table to lookup the
full string of an interned symbol. The cleanest way to implement this
would be to pass a pointer to the field layout table when starting the
task. This allows multiple Arret binaries (e.g. different libraries and
plugins) to live in a single process without interfering with each
other.

However, this causes an additional annoyance for our `PartialEq` and
`Hash` implementations. We now need access to the `Heap` to lookup the
record's class ID. This prevents us from implementing a simple
`PartialEq`/`Hash` across all boxed values.

This instead introduces `PartialEqInHeap` and `HashInHeap` which is the
same as the Rust traits instead they accept a `Heap`. There's a default
implementation for boxed values that implement `PartialEq` and `Hash`
which means most values are still comparable. Notably, however, `Any`,
`Pair`, `List` and `Vector` now require explicit usage of the new
traits.