Allow records to store data
This allows records to store an arbitrary amount of `[u8]` data.
Depending on the size of the data this will either be stored as a 16
byte inline box, 32 byte inline box or as external data pointer.

This is modelled after how `Str` works as they're both essentially doing
the same thing. One difference is that `Str` tries to be flexible for
various sizes of `Arc` while effectively hardcodes 32bit and 64bit
layouts.

This moves assertions for unsupported behaviour (equality, hashing,
garbage collection) in to where they occur. This will allow us to
incrementally introduce users of records (such as closures) without
having to implement everything at once.