Remove support for jumbo const records
Constant records weren't respecting `MAX_HEAP_INLINE_BYTES` and instead
allowed records up to 254 bytes long. The idea was we weren't limited by
the size of the box and instead could store as much as we could record
in `inline_byte_length`.

However, this presents difficulty for the record field accessors - they
don't know if they're dealing with a constant or a heap value without
checking. There is less complexity and code bloat if records are always
represented the same way. For that reason we should limit constant
records to the same representations as heap records.