Remove support for callback record fields
This was never used due to `specific_abi_type` not selecting it.
However, there's no way to implement this without supporting conversion
from callbacks to fun thunks. This is required to box the return value
from the field accessor at a minimum.

One of the messes that etaoins/llambda got it to was being eager to
build layer after layer of thunks to convert function values. I had
decided to only allow one layer of conversion (callback => thunk) in
Arret to avoid that. This means we should only use fun thunk values in
records and let specialisation optimise for us when it can.