Multithread Valgrind tests
Valgrind is inherently single threaded and uses user-space locks to
prevent multiple threads from running concurrently. This means we're
wasting half of the CPU available on our CI servers.

Work around this by determining the number of host CPUs and launching
that number of Valgrind processes. Each process takes a portion of the
work based on environment variables passed in to identify the worker.