Use precise alignments in record data
We currently treat all record data as 8 byte aligned with the assumption
that this is the worse case alignment. This has two disadvantages:

1. If an architecture or data type requires alignment of 16 bytes or
   more we'll just assert at compile time.

2. If record data requires less than 8 bytes of alignment we'll waste
   memory on padding.

However, we already need to dynamically calculate the record data's size
so we also know its alignment. This replaces the places we tracked
`data_len` to instead track the `alloc::Layout` which contains the
alignment information. If the layout is incompatible with our inline
data (size or align) we'll store it externally

This also adds a `RecordData` helper object that tracks the allocation
along with its layout. This allows us to reuse this MIR's allocation in
the final external record.