Track MIR values' known record conses
This is analogous to how we track `possible_type_tags` except we only
support a single known record cons at a time. This is to balance the
overhead and complexity of tracking multiple record conses versus the
benefit. While type tags can be very compactly encoded, intersected,
etc. because they can fit inside a bitmask there can be an arbitrary
number of record conses.

The known record cons is conditional on the value actually being a
record. This allows us to track the record cons through unions with
non-record values. This can allow for some further optimisations of
unions with a single record value but means additional care is needed to
not assume any value with a record cons is actually a record.

This is currently only used to optimise our typred code for the edge
case where we know the record cons but not if the value is a record.
However, this should be much more useful for building inline equality
tests between two values of a known record cons.