Consider MIR box casts as free for costing
The old `RegCast` category used to also cover integer type conversions
which could theoretically involve extending an integer value. For that
reason it was given the minimum possible cost of 1.

However, since 32bit support has been dropped the category only contains
`CastBoxed`. This is simply reinterpreting a pointer so it should be
free like it's const equivalent (`ConstCastBoxed`).

This is motivated by some dubious costing for record operations that
need to cast a value when loading each field. While this is free the
`RegCast`s were inflating the calculated cost of the ops for the
purposes of inlining.