Implement (some?)
This is actually more similar to Clojure's `(some)` which finds the
first truth-y value returned by the mapper function. Clojure's `(some?)`
basically casts a truthy value to a boolean. Seeing as we don't have the
concept of truthiness a Clojure-style `(some?)` doesn't make sense and
Clojure's `(some)` becomes a predicate.