Fix TypeTagSet from an intersect ty ref
Due to a copy-and-paste bug we were starting out intersection type tag
set with the empty set. This is then `&`ed with every member which will
always produce the empty set. This caused our MIR to determine that
certain values value no possible types which is obviously nonsense. The
empty intersection is actually all possible types so we should start
with `TypeTagSet::all()`.

This adds both a low-level test for the specific buggy function and a
higher level test for an observed breakage.