Add runtime SharedStr implementation
This replaces our boxed string's `Arc<str>` with a type with a known FFI
layout. This will enable codegen to create non-inline boxed strings.

The implementation is taken from etaoins/llambda's `SharedByteArray.cpp`
and wrapped in a safe Rust type. It has a side benefit of allowing our
non-inline strings to be stored in a 16 byte box.