Add helper for building synthetic funs
Because we build an entire inferred AST by hand for each synthetic fun
it's very verbose and prone to error.

This factors our the common patterns with our 4 existing synthetic funs
in to a new `wrap_poly_expr_in_arret_fun`. It does the heavy lifting of
building the AST given a parameter, body expression and some types.