Fix recursive funs with polymorphic purity
MIR has a special function called `fun_app_purities` which attempts to
figure out an application's purity based on application's type args and
the environment's type args. It does not account for the mono type args
in the function context. This falls apart in a case like this:

```
(defn times #{[->_ ->!]} ([i Int] [body (->_ ())]) ->_ ()
  (when (> i 0)
    (body)
    (times (- i 1) body))) ; <== THIS APPLICATION
```

In this case:

1. This is a top-level function so it has no environment type args.

2. The applications type args are `->_` == `->_` (recursively defined)

3. The function context will have a type arg set for `->_`. This will
   typically be the upper bound of `->!`.

When trying to determine the application's purity we get caught in a
loop trying to resolve `->_` to `->_`. We don't account for the function
context at all so we never end up with `->!`.

What we really want is to have a three input operation where we subst
the application type args using the context and then add it to the
environment. That would correctly resolve `->_!` during the application.

This exact operation already exists as `merge_apply_ty_args_into_scope`
which was already used for inline function application. For some reason
this wasn't substing purity but that's easily fixed.

This also splits out the purity part of merging in to
`merge_apply_purity_into_scope`. This is an optimisation to avoid also
merging the ty args in situations where they would just be discarded.