Add specific type error for return values
This adds a specific type inference error for when a value doesn't match
the expected return value. This can be helpful when it isn't clear that
an expression is actually the terminal expression in a function body
(due to macros, nesting, etc) or if the return type is visually far away
from the error site. It looks the same as the existing type error but
adds a secondary span pointing to the return type declaration.

This replaces our type checker's `required_type` with a new `result_use`
enum that describes how the expression's result will be used. This is
also used to make our dead code elimination slightly smarter by
recognising that the terminal expression in an unused `(do)` is also
unused. This doesn't seem to occur often in practice.

The original motivation for this change was to allow our type checker to
enforce that `(recur)` always occurs in a tail position. However, the
nicer error messages are worth this machinery by themselves.