Fix typeck dead code elimination for applications
We were incorrectly checking the purity of the return type, not the
function itself. For non-function return values this would fall back to
thinking the expression can have a side effect which is suboptimal but
safe. However, in the case of an impure function returning a pure
function this could falsely eliminate code with side effects.

This also removes a dead comment referring to when we ran DCE for an
expression before inferring it.