Remove duplicate alloc ops
This adds a new MIR optimisation pass to remove duplicate alloc ops.

This only works on simple boxing of scalar native values. This is
because we use a hacky `OpKind::Alias` op to create aliases to existing
LLVM values. This means that if we were de-duplicating, e.g. a list
containing an integer we wouldn't recognise duplicate pair allocations
because they would all come from different aliases to the same integer
reg.

The proper solution for this is to re-write all reg references like
Llambda does. However, this is quite complex and there's no
infrastructure we can immediately re-use.