Make TypeTagSet from ty refs handle polys
Right now if `TypeTagSet::from` is passed a poly type var it will create
a `TypeTagSet::all`. This is technically correct but suboptimal for e.g.
a poly type var bounded by `Num`.

We're about to tweak how we pick function signatures to help with tail
recursion. We'll need this to choose better ABIs for polymorphic
functions.