Return immediately after `(recur)` when possible
LLVM's tail call optimisation requires that we `ret` immediately after
`call`. Potentially things like `phi`ing through a conditional could
break that.

Because `(recur)` is type checked to only appear in a return value
position we can build a return directly to minimise the instructions
appearing between the call and the return. This only works if we're not
inlined inside a different function.