Don't (recur) during MIR evaluation
When doing constant evaluation (e.g. in REPL) we can't properly support
tail recursion within our MIR evaluator. Rust does not support tail
recursion and even if it did it would be a monumental effort to
structure the compiler in a way that expression evaluation always tail
recurred.

We can do exactly what we do with Rust functions and build a thunk and
then call in to that. Because the thunk is LLVM generated Arret code
it's already required to be tail recursive. This isn't always upheld at
the moment but any time we fail to tail recurse is a bug that will be
fixed.