Initial support for dumping MIR
Currently it's quite unpleasant to debug situations where we generate
illegal MIR that causes our codegen to crash. Up until this point I've
been writing ad-hoc printers that iterate over ops and use their `Debug`
implementation. In addition to being duplicate effort the output is
verbose and hard to follow.

This adds initial support for dumping MIR through two mechanisms:

- A special `.mir` file extension from the driver. The other file
  extensions were documented at the same time.

- An `ARRET_DUMP_MIR` environment variable that's analogous to
  `ARRET_DUMP_LLVM`

The output format is inspired by Rust but has parallels to LLVM IR due
to being lower level and SSA. Not all ops are pretty printed at the
moment; we'll fall back to `Debug` if we don't recognise the ops. The
idea is that we can incrementally support pretty printing as their ops
appear in MIR. One big omission at the moment is `Cond` but there's some
infrastructure in place for printing indented branches that is intended
for use with `Cond`.